Implement calculation for 'corrected sample standard deviation'
Writing the logic for the 'corrected sample standard deviation' based on the Ticks of the TimeSpan quickly results in OverflowExceptions for cycle times expressed in days. Writing the calculation using TimeSpan.TotalDays resulted in calculations off by more than 1ms when cycle times where expressed in the order of miliseconds. As such settled for now on using TimeSpan.TotalHours for the calculation to keep the limited precision of the calculation within 1ms of the test oracle.
However this level of accuracy might be immensely overkill? Task durations where humans are involved in doing it are typically not at the resolution of milliseconds.

Also improves PropertyChangedEventTracker to work as expected when other property change events have been recorded.