Create TimeTillCompletionEstimator
Based on random sampling of a collection of InputMetrics, the TimeTillCompletionEstimator will determine an estimate on how many working days are needed to complete the given number of working items.

Because 0 throughput values are allowed on InputMetrics, a maximum number of iteration is recommended to be provided to ensure that the estimate will always be given within a reasonable amount of time.
Failing to complete the work estimate within the given iteration limit, will mark the estimate as indeterminate.