Refactor TimeTillCompletionEstimator to depend on the concept of a Project
This refactor is a preparation to defining a roadmap that consists out of multiple projects, and being able to estimate the time it takes to complete a Roadmap of Projects.
Behaviorally from the perspective of the user, nothing changed. The user still provides a number of work items to be completed for a project, of which the time to complete it is forecast by the app.