Introduce RoadmapConfiguratorControl as dedicated UserControl for Roadmap configuration
HACK: In order to keep good visual outlining, I fixed the width of the MainWindow label section and copied that same width into the RoadmapConfiguratorControl too. Not too pretty solution, but gets the job done.
Not sure if I would like to solve this through Styling, a constant class or some UI concept that represents that label/input section thingy. Therefore leaving this for a later moment, when the right solution pops to my mind. (if ever)