TimeTillCompletionEstimator calculates both the estimate for the whole roadmap, alongside projects
When multiple projects are being worked on, individual projects could be finished at any point. But the time it takes to finish the roadmap is the time that all projects have finished, which is typically longer. This should be captured as well, as typically one is in the end (also or only) interested when the whole roadmap is completed.