Dynamically create the required columns DataGrid showing estimations
This makes it possible to dynamically adjust the shown columns based on the projects configured for the roadmap being estimated.

Currently implemented using a property of DataGridColumns on MainWindowViewModel, which isn't the cleanest solution. ViewModels typically shouldn't have dependencies to View-style classes. First want to make sure that dynamic creation of columns can be fully implemented outside the MainWindowViewModel class. And if so, probably will refactor the code to push it into the MainWindow class.