Refactor how work estimations end up in DataGrid table
Simplified the design the allow for dynamic columns in the DataGrid showing the work estimations, through the use of the new TimeTillCompletionEstimationsCollectionToDataTableConverter. This eliminates the need for the DataGridAttachedProperties class and the datagrid properties managing code in MainWindowViewModel, which have now have a better designed separation of responsibilities.
MainWindowViewModel now exposes the estimations straight as TimeTillCompletionEstimationsCollection.
TimeTillCompletionEstimationsCollectionToDataTableConverter facilitates the same behavior as before, but can be extended in a later commits to become an implementation of IMultiValueConverter in order to dynamically control which data content is shown in the table.

The DataGrid still only shows the estimation of work for only the roadmap, and the export similarly still only exports the work for the roadmap.