Split off 'Application' assembly from Domain and created corresponding test project
The concept of 'IRandomNumberGenerator' residing in the Domain layer of the 'clean architecture' is debatable.
Considering that the implementation for that interface could be considered an 'interface adapter' layer object and that the Application business logic is the only dependency on that interface,
everything depending on IRandomNumberGenerator has been pulled into the new Application layer of the solution.
A new unit tests project has been created for the Application layer objects.