Defined Project- and RoadmapConfiguration classes
These classes serve to make a distinction between the classes used during simulation (which track work being done during the simulation) and the classes that represent the specification of the work to be done (which ought to be immutable).
This can help simplify edge-cases and business logic.
This is also the first step to move some business logic out of the RoadmapConfigurationViewModel, and into the 'Application' architectural layer of the solution.